Generalize the path-find mechanism
I have tried to generalize the entire procedure of path-find.
First, I have created an abstract class called GraphNode which exposes three functions provided to the path-find environment. Room now extends this class, so it can be used withing the path-find environment.

Furthermore, another I have create another abstract class called PathFinder which requires only one function which is findPath.

However, the algorithm should not accept paths which traverse wrong rooms (closed doors, flying unconnected rooms, etc.) so the findPath function accepts a function pointer. This function pointer is a function which is used to determine if there is a valid connection between a From and To nodes.