Handle TwoHanded weapons
Basically if a weapon is a TwoHanded weapon, it occupies two slots!
IDEA: Add a flag or a table or whatever which state if a model Requires (that's the deal) multiple body-parts!
Basically, this way for instance, an exoscheleton which occupies all the body pats of a generic living-beeing can be created (it is quite interesting...).