Implement a first prototype of the new skill system
Basically there are inside the mud four major categories of modifiers (ability, status, combat and knowledge). A new table associate a skill to a modifier and a required rank. Whenever the skill reaches the specified rank, the modifier is activated. An example: when Material Gathering reaches the rank 470, the modifier GatherMetal is activated, allowing the character to gather metals.