Register mud datas and structures inside the lua_script file
I've moved the process which registers all the datas and structures of the Mud inside a single file. This way is much more cleaner. A lot of code was useless, for instance all the functions which were wrapping vectors with LuaRefs. Now since the vectors can be registerd is easier.